Add files via upload
Changed some of the priors and also the way the fits are plotted now.
The results of scipy.optimize.minimize are highly dependent on the initial values I put in and on my priors.
I've fiddled a bit more with the priors, e.g. I lowered the maximum value for the PL normalization.